requirements.txt
asgiref==3.9.1
boto3==1.40.40
botocore==1.40.40
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
django-admin-thumbnails==0.2.9
django-allauth==65.11.2
django-session-timeout==0.1.0
django-storages==1.14.6
idna==3.10
jmespath==1.0.1
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
